+++ src/constants/translations/ua/common.json
@@ -64,5 +64,9 @@
     "french": "Французька",
     "spanish": "Іспанська",
     "arabic": "Арабська"
+  },
+  "categoryStep": {
+    "title": "Будь ласка, виберіть основні теми на основі категорії. Ви можете додати інші пізніше",
+    "buttonText": "Додайте ще один предмет"
   }
 }
